AI_Medical_Chatbot/requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_experimental
google-generativeai
langchain-google-genai
-e .